refactor: auth provider confings
