+++ src/routes/OpenidAuthRoutes.ts
@@ -1,6 +1,7 @@
+import SinunaSettings from "../providers/sinuna/Sinuna.config";
 import { generateRequestHandlers } from "../utils/route-utils";
 
-const defaultAuthProviderIdent = "sinuna";
+const defaultAuthProviderIdent = SinunaSettings.ident;
 const operationPrefix = "OpenId";
 const operationNames = ["LoginRequest", "AuthenticateResponse", "AuthTokenRequest", "LogoutRequest", "UserInfoRequest", "AuthorizeRequest"];
