+++ src/routes/Saml2AuthRoutes.ts
@@ -1,6 +1,7 @@
+import SuomiFISettings from "../providers/suomifi/SuomiFI.config";
 import { generateRequestHandlers } from "../utils/route-utils";
 
-const defaultAuthProviderIdent = "suomifi";
+const defaultAuthProviderIdent = SuomiFISettings.ident;
 const operationPrefix = "Saml2";
 const operationNames = ["LoginRequest", "AuthenticateResponse", "LogoutRequest", "UserInfoRequest"];
